refactor: changing folder dashboard to chart
