+++ frontend/src/display_content/model/chart/line/LineChart.js
