+++ frontend/src/display_content/model/chart/line/WideLineChart.js
